userInteractive
