+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@tippyjs/react": "^4.2.6",
+    "animate.css": "^4.1.1",
     "axios": "^1.3.4",
     "classnames": "^2.3.2",
     "react": "^18.2.0",
