+++ src/components/Foryou/ListContent/HeaderContainer/HeaderContainer.module.scss
@@ -1,0 +1,50 @@
+.header-container {
+    display: flex;
+    align-items: flex-start;
+    justify-content: space-between;
+    .info-container {
+        .wrap {
+            cursor: pointer;
+            .nickname {
+                font-weight: 700;
+                font-size: 17px;
+                line-height: 25px;
+                margin-right: 4px;
+            }
+            .name {
+                font-size: 13px;
+            }
+            &:hover .nickname {
+                text-decoration: underline;
+            }
+        }
+        .status {
+            font-weight: 300;
+            font-size: 15px;
+            .hash-tag {
+                font-weight: 500;
+                cursor: pointer;
+                &:hover {
+                    text-decoration: underline;
+                }
+            }
+        }
+        .music {
+            display: inline-block;
+            margin: 4px 0px 12px;
+            font-size: 15px;
+            font-weight: 500;
+            cursor: pointer;
+            &:hover {
+                text-decoration: underline;
+            }
+            i {
+                margin-right: 5px;
+            }
+        }
+    }
+    button {
+        margin-top: 9px;
+        margin-right: -5px;
+    }
+}
