+++ src/components/Foryou/ListContent/VideoContent/VideoContent.module.scss
@@ -1,0 +1,158 @@
+.video-wrapper {
+    display: flex;
+    justify-content: flex-start;
+    align-items: flex-end;
+    .wrapper {
+        margin-top: 5px;
+        margin-right: 20px;
+        width: 280px;
+        height: 500px;
+        position: relative;
+        &:hover {
+            .play-button,
+            .volume-wrapper,
+            .timeline-wrapper {
+                opacity: 1;
+                visibility: visible;
+            }
+        }
+        video {
+            width: 100%;
+            height: 100%;
+            object-fit: cover;
+            border-radius: 8px;
+        }
+        .play-button {
+            opacity: 0;
+            visibility: hidden;
+            color: white;
+            position: absolute;
+            bottom: 7.5%;
+            left: 9%;
+            cursor: pointer;
+        }
+        .volume-wrapper {
+            opacity: 0;
+            visibility: hidden;
+            height: 45px;
+            width: 45px;
+            position: absolute;
+            bottom: 5%;
+            right: 4%;
+            cursor: pointer;
+            &:hover .fill-bar {
+                display: block;
+            }
+            &:hover .volume-bar {
+                display: block;
+            }
+            .volume-button {
+                color: white;
+                position: absolute;
+                top: 50%;
+                left: 50%;
+                transform: translate(-50%, -50%);
+            }
+            .volume-bar {
+                display: none;
+                padding: 11px 7px;
+                width: 65px;
+                border-radius: 40px;
+                background-color: rgba(0, 0, 0, 0.387);
+
+                position: absolute;
+                top: -40px;
+                right: -10px;
+                transform: rotate(270deg);
+                cursor: pointer;
+
+                -webkit-appearance: none;
+
+                &::-webkit-slider-runnable-track {
+                    height: 2px;
+                    background: rgba(235, 233, 233, 0.323);
+                    border-radius: 3px;
+                    z-index: 10;
+                }
+
+                &::-webkit-slider-thumb {
+                    -webkit-appearance: none;
+                    height: 14px;
+                    width: 14px;
+                    border-radius: 50%;
+                    background: white;
+                    margin-top: -6px;
+                    z-index: 10;
+                }
+            }
+            .fill-bar {
+                display: none;
+                width: 2px;
+                height: 50px;
+                background-color: white;
+                position: absolute;
+                bottom: 48px;
+                right: 47%;
+                z-index: 2;
+            }
+        }
+        .timeline-wrapper {
+            opacity: 0;
+            visibility: hidden;
+            width: 268px;
+            padding: 10px 5px;
+
+            position: absolute;
+            bottom: 2px;
+            left: 3%;
+
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            &:hover {
+                .timeline {
+                    &::-webkit-slider-runnable-track {
+                        height: 4px;
+                    }
+                    &::-webkit-slider-thumb {
+                        background: white;
+                    }
+                }
+                .fill-bar-timeline {
+                    height: 4px;
+                }
+            }
+            .timeline {
+                width: 185px;
+                -webkit-appearance: none;
+                background: rgba(255, 255, 255, 0.34);
+                cursor: pointer;
+                &::-webkit-slider-runnable-track {
+                    height: 2px;
+                    border-radius: 3px;
+                    z-index: 10;
+                }
+                &::-webkit-slider-thumb {
+                    -webkit-appearance: none;
+                    height: 14px;
+                    width: 14px;
+                    border-radius: 50%;
+                    // background: white;
+                    margin-top: -6px;
+                    z-index: 10;
+                }
+            }
+            .running-time {
+                font-size: 10px;
+                color: white;
+            }
+            .fill-bar-timeline {
+                width: 0px;
+                height: 2px;
+                background-color: white;
+                position: absolute;
+                z-index: 2;
+            }
+        }
+    }
+}
